% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_odd_streams.R
\name{find_odd_streams}
\alias{find_odd_streams}
\title{Detect outlying series within a collection of  sreaming time series}
\usage{
find_odd_streams(test_stream, pc, threshold, update_threshold = TRUE,
  update_threshold_freq)
}
\arguments{
\item{test_stream}{A multivariate streaming time series data set}

\item{pc}{The pricipal component attributes  returned by \code{\link{get_pc_space}}}

\item{threshold}{A threshold value to determine outlying series. A value returned by
\code{\link{set_outlier_threshold}}}

\item{update_threshold}{If TRUE, the threshold value to determine outlying series is updated.
The default value is set to TRUE}

\item{update_threshold}{A numerical value to indicated how often the threshold should be updated.
 (After how many windows
it should be updated)}
}
\value{
The indices of the outlying series in each window. For each window a plot is also produced on the current
graphic device
}
\description{
This function detect outlying series within a collection of streaming time series. A sliding window
is used to handle straming data. In the precence of concept drift, the forecast boundary for the system's typical
behaviour can be updated periodically.
}
\seealso{
\code{\link{get_pc_space}}, \code{\link{set_outlier_threshold}}
}
