% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc_space.R
\name{plotpc}
\alias{plotpc}
\title{Plot a two dimensional feature space on the current graphics device.}
\usage{
plotpc(pc_pcnorm)
}
\arguments{
\item{pc_pcnorm}{The scores of the first two pricipal components returned by \code{\link{get_pc_space}}}
}
\value{
A graphical representation of the two dimensional feature space will be produced on the current graphic
device.
}
\description{
Plot a two dimensional feature space on the current graphics device using the first two
pricipal component returned by \code{\link{get_pc_space}}
}
\examples{
#Generate training dataset
set.seed(123)
nobs = 500
nts = 50
train_data <- ts(apply(matrix(ncol = nts, nrow = nobs), 2, function(nobs){10 + rnorm(nobs, 0, 3)}))
features <- extract_tsfeatures(train_data)
pc <- get_pc_space(features)
plotpc(pc$pcnorm)
}
\seealso{
\code{\link{find_odd_streams}},  \code{\link{extract_tsfeatures}}, \code{\link{get_pc_space}},
\code{\link{set_outlier_threshold}}
}
