% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_outlier_threshold.R
\name{set_outlier_threshold}
\alias{set_outlier_threshold}
\title{Set a threshold for outlier detection}
\usage{
set_outlier_threshold(pc_pcnorm, p_rate = 0.001, trials = 500)
}
\arguments{
\item{pc_pcnorm}{The scores of the first two pricipal components returned by \code{\link{get_pc_space}}}

\item{p_rate}{False positive rate. Default value is set to 0.001}

\item{trials}{Number of trials to generate the extreme value distirbution. Default value is set to 500.}
}
\value{
Returns a threshold to determine outlying series in the next window  consists with a collection of
time series.
}
\description{
This function forecasts a boundary for the typical behaviour using a representative sample
of the typical behaviour of a given system. An approach based on extreme value theory is used for this boundary
prediction process.
}
\references{
Clifton, D. A., Hugueny, S., & Tarassenko, L. (2011). Novelty detection with multivariate extreme value statistics.
Journal of signal processing systems, 65 (3),371-389.
}
\seealso{
\code{\link{get_pc_space}}
}
