% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anomalous_stream}
\alias{anomalous_stream}
\title{Multivariate timeseries dataset with an anomalous event.}
\format{A data frame with  640 series each with 1459 time points.}
\usage{
anomalous_stream
}
\description{
A mutivariate time series dataset with some anomalous series. These time series are with noisy signals.
}
\keyword{datasets}
