% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc_space.R
\name{get_pc_space}
\alias{get_pc_space}
\title{Define a feature space using the PCA components of the feature matrix}
\usage{
get_pc_space(features, robust = TRUE)
}
\arguments{
\item{features}{Feature matrix returned by \code{\link{extract_tsfeatures}}}

\item{robust}{If TRUE, a robust PCA will be used on the feature matrix}
}
\value{
It returns a list with class 'pcattributes' containing the following components:
   \item{pcnorm}{The scores of the firt two pricipal components}
   \item{center, scale}{The centering and scaling used}
   \item{rotation}{the matrix of variable loadings (i.e., a matrix whose columns contain the eigenvectors).
                   The function \code{princomp} returns this in the element loadings.}
}
\description{
Define a two dimensional feature space using the first two principal components generated from
the fetures matrix returned by \code{extract_tsfeatures}
}
\examples{
#Generate training dataset
set.seed(123)
nobs = 500
nts = 50
train_data <- ts(apply(matrix(ncol = nts, nrow = nobs), 2, function(nobs){10 + rnorm(nobs, 0, 3)}))
features <- extract_tsfeatures(train_data)
pc <- get_pc_space(features)

}
\seealso{
\code{\link[pcaPP]{PCAproj}}, \code{\link[stats]{prcomp}}, \code{\link{find_odd_streams}},
\code{\link{extract_tsfeatures}}, \code{\link{set_outlier_threshold}}, \code{\link{plotpc}}
}
