% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc_space.R
\name{plotpc}
\alias{plotpc}
\title{Plot a two dimensional feature space on the current graphics device.}
\usage{
plotpc(pc_pcnorm)
}
\arguments{
\item{pc_pcnorm}{The scores of the first two pricipal components returned by \code{\link{get_pc_space}}}
}
\value{
A graphical representation of the two dimensional feature space will be produced on the current graphic
device.
}
\description{
Plot a two dimensional feature space on the current graphics device using the first two
pricipal component returned by \code{\link{get_pc_space}}
}
\seealso{
\code{\link{get_pc_space}}
}
